fix: added swal alerts

+++ src/screens/Detailpage.jsx
@@ -3,6 +3,7 @@ import url from '../BackendURL';
 import { useNavigate, useParams } from 'react-router';
 import axios from 'axios';
 import List from '../components/lists/List';
+import Swal from 'sweetalert2';
 
 function Detailpage() {
   const navigate = useNavigate();
@@ -27,11 +28,26 @@ function Detailpage() {
    * deletes a list by its id
    */
   const deleteList = async () => {
-    await axios.delete(url + '/api/lists/' + list.id)
-      .then((res) => {
-        console.log(res.data);
-        navigate('/');
-      });
+    Swal.fire({
+      position: 'center',
+      icon: 'warning',
+      iconColor: '#ff6a6a',
+      title: 'Bist du sicher, dass die gesamte Liste inkl. aller Todos gelöscht werden soll',
+      showConfirmButton: true,
+      confirmButtonText: 'Ja',
+      confirmButtonColor: '#bef983',
+      showCancelButton: true,
+      cancelButtonText: 'Nein',
+      cancelButtonColor: '#ff6a6a',
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        await axios.delete(url + '/api/lists/' + list.id)
+          .then((res) => {
+            console.log(res.data);
+            navigate('/');
+          });
+      }
+    });
   }
 
   return (
